Update the workshop to obtain examples and setup_demo.sh from a released branch.
Using a release branch can shield against any breaking changes we could introduce between versions (since TFX is still at 0.x). And we can cherry-pick specific fixes applicable to a version if necessary.

The downside of this is that fixes to setup_demo.sh to master needs to manually cherry picked.

PiperOrigin-RevId: 257476236