PR #477: Pass exec_properties to BaseDriver's _prepare_output_artifacts
Please approve this CL. It will be submitted automatically, and its GitHub pull request will be marked as merged.

Imported from GitHub PR #477

It's currently very difficult to override the `_prepare_output_artifacts` function in the `BaseDriver`.
This function currently receive only basic pipeline and component information as well as the `output_dict`.
That means you cannot have external parameters in your custom logic.

This PR passes the `exec_properties` dictionary as a parameter, in a similar way to what is done in `resolve_input_artifacts`.
It also rename `_prepare_output_artifacts`  to `prepare_output_artifacts` to make it "public" according to Python standard

Copybara import of the project:

  - 0f22492643f9eee1a4cbe3fca96bf41f135b6473 Pass exec_properties to driver's _prepare_output_artifacts by Martin Laurent <mlaurent@spotify.com>
  - c1224c31c2151603a067d809ebe5b49af4edf4d7 pylint and make the method public by Martin Laurent <mlaurent@spotify.com>
  - 49d4a14c7e0a3ed81e8663e0f909e5ff85400002 Merge c1224c31c2151603a067d809ebe5b49af4edf4d7 into 188df... by Martin Laurent <mlaurent@spotify.com>

PiperOrigin-RevId: 267454254