Refactor KubeflowDagRunner to directly manipulating dsl.Pipeline object. Providing flexibility for pipeline parameterization.
Currently, pipeline parameter specification in KubeflowDagRunner is done by the signature of pipeline function, which is immutable. In order to support TFX pipeline parameters, we need to manipulate dsl.PipelineParam associated with a pipeline object programmatically. This change exposes the dsl.Pipeline object in KubeflowDagRunner so that further change can be made to support pipeline parameterization.

PiperOrigin-RevId: 269386565