Add PVC support for InfraValidator KubernetesRunner.
PVC support is needed for on-prem users with custom volume/storage attached.

Whether pipeline data root is a mounted volume or not is automatically detected. Detection is done by checking every volume mount whether it contains the model_path in its mount_path.

Mounted volume must be a PersistentVolumeClaim type in the same namespace, and its access mode should be a ReadWriteMany mode.

PiperOrigin-RevId: 305055491