Update ci-test.yml
Fixes the case when there are added files and modified files together.

Because there is no new line at the end of the file list, the last entry of the first list was stuck to the first entry of the second list. Added a new line with `echo` to resolve this.