Streams output of `skaffold` to improve usability.
It takes several minutes to build an image while creating a pipeline with KFP. However, currently there is no output shown and it is hard to know whether it is stuck somewhere. With this output streaming, users can know the progress well.

Also updated not to use deprecated `.IMAGE_NAME` in envTemplate, and deleted unused files.

PiperOrigin-RevId: 325138187