Restrict python version used in Airflow Workshop to < 3.8
tfx==21.1 requires ml-metadata>=0.21.2,<0.22 [1].
There exist no ml-metadata published packages in that version range that have been built with cpython version 3.8; the only packages that match the cp38 tag are 0.24 packages. [2] Therefore, we must disallow users to run this tutorial with python >3.7.

This CL also bumps airflow to incorporate a fix to airflow [3] that will unbreak the tutorial because of this reported issue.[4]

[1] https://github.com/tensorflow/tfx/blob/r0.21.1/tfx/dependencies.py#L33
[2] https://pypi.org/simple/ml-metadata/
[3] https://github.com/apache/airflow/pull/8230
[4] #1497

PiperOrigin-RevId: 332732122