Expand the Penguin example to include the option to define the model using Flax in addition to Keras.
The Flax support in TFX is experimental.

Only the Trainer component of the pipeline needs to be changed for Flax.

In order to highlight that there are very few changes compared to a Keras pipeline,
I have parameterized the Penguin example with the option to use either Keras or Flax for the model. For `penguin_pipeline_local.py` the option is passed through a
flag `model_framework` with the values `keras` (default) or `flax_experimental`.

PiperOrigin-RevId: 366940246