Adds DagRunnerPatcher for communication between CLI and users' DagRunner scripts.
These patcher classes "decorate" the `run` function of the DagRunners and it can intercept arguments before/after the run() execution, and it can even skip the real run() execution. For example, it can be used to read the name of the pipeline or root directory without really executing run() function in the dag runner script file.

KubeflowDagRunnerPatcher will read/update the image and the kfp output path as needed.

Also adds APIs to access config object of the DagRunner.

PiperOrigin-RevId: 369602168