Fixes out-dated CLI e2e test.
We don't make pipeline package file when creating pipelines any more.

Also we need to invoke CLI in a separate process. Because when we import the file for the second time, the runner will not be actually called because the import doesn't happen. So replaced runner.invoke() with a shared utility which uses subprocess.run().

PiperOrigin-RevId: 370036319