Resolver can use ResolverFunction as an input resolution logic.
ResolverFunction is compiled to the same ResolverConfig.ResolverSteps IR.
This is because we restrict ResolverOp topology to be linear. This
restriction may be relaxed in the future when ResolverOp should take more
than one argument (i.e. becomes binary+ operator), where new graph
topology based IR is needed.

Also adds @resolver decorator for a Resolver factory.

PiperOrigin-RevId: 370621097