Fixes issues in https://github.com/tensorflow/tfx/pull/3983.
1) Note on the cherry pick in https://github.com/tensorflow/tfx/pull/3983.
https://github.com/tensorflow/tfx/commit/6f073c9f7570d6e6ab29ec1fcc14e31bce506222#diff-e1fd4c84c29238e5c1ab64975d2032afd8d5374aeea170347448eb28f9ab37bd

It seems like the import statement `from tfx.dsl.components.base import node_registry` should be deleted.
Because `node_registry.py` doesn't exist in 1.0.0 branch, it will result in a import error, and it seems like the import can be deleted.

2) https://github.com/tensorflow/tfx/pull/3983 didn't fix already
deleted code after 1.0.0 branch cut. get_all() is not needed because
inputs and outputs are just a Dict.