Delete container images including not tagged one in test clean-up.
We only deleted image with the current tag, and it leaks untagged images which is obsolete versions of the tag. This PR will list all digests and delete one by one. The deletion should happen from child to parent order, so we delete target images first and then delete base images.

Also introduce a dedicated module for container image management.

PiperOrigin-RevId: 395031833