Reload `pkg_resources` when not running in colab.
Deep in the TF imports, `google-api-python-client` uses `pkg_resources` to do some introspection, however it uses a cached copy of `sys.path` that's generated at import-time. In this notebook, an earlier cell is running `pip install`, which changes the installed packages, invalidating the package cache and causing errors.

This change re-imports the module, refreshing the cache, so that the notebook can proceed when executed in our doc generation infrastructure.

PiperOrigin-RevId: 398677602