LatestBlessedModelStrategy should gracefully handle no inputs.
In our code example [1], we use resolver with LatestBlessedModelStrategy and expect the first pipeline run without any previously blessed models should be okay to run the resolver. However, old implementation simply returned None which is treated as an input resolution error. The reason this was not captured from the test was that Launcher has been ignoring the resolver node errors.

[1] https://www.tensorflow.org/tfx/guide/evaluator

PiperOrigin-RevId: 436341676