Refactor inputs_utils.resolve_input_artifacts.
This also changes min_count behavior; Previously we filter out dicts whose
min_count is not met and raise an error if the resolved result is empty.
However empty list is not really an error, especially if ForEach + Cond is
mixed and all ForEach slice conditional evaluated to false (results in empty
list), it is definitely not an error.

For min_count it is better we directly raises an error for any insufficient
input dict encounter. This change does contain a side effect but impact would
be zero because currently there isn't any use cases where there is multiple
resolved input dicts but artifact length per input key is different.

PiperOrigin-RevId: 455551308