Fixes a bug in the Artifact attribute setter that was causing the corresponding getter not to return a value for properties of type JSON_VALUE.
When the setter was invoked on a JSON value attribute, we previously were not storing the value in the artifact's properties dict (we only stored it in the cached JSON value properties). This was causing the getter to fail to return a value for JSON values, as the getter first checks that the property name exists in the artifact properties dict before looking up the value.

PiperOrigin-RevId: 459449777