Temporarily fix the dependency of TF-serving-api==2.9 to allow TF-2.9.0rc2.
The GCP DLVM container image for TF 2.9 contains TF-2.9.0rc2 and TF-serving-api 2.9.0 which is problematic because TF-serving-api==2.9 requires a regular version of TF 2.9, but we cannot re-install TF in DLVM images to avoid messing up subtle optimizations. This is a temporary workaround until GCP DLVM image contains regular TF version.

PiperOrigin-RevId: 459680430