Retries TF serving API import if the import fails due to a proto change.
Tensorflow [moved the location of error code definitions](https://github.com/tensorflow/tensorflow/commit/dc2bfae69be58eea63927ffa5900d70b349d1638), but it breaks the code if the version of TF doesn't match with TF serving API.

Because TF serving API doesn't provide nightly packages, we don't have a way to run a test that depends on TF serving API when we need to use TF-nightly.

This code is a hack to make the moved code work. The only symbol that is missing in the new location is '_CODE', so we manually inject it to make the import successful.

PiperOrigin-RevId: 479004136