Fix OutputSpec IR compilation for OutputChannel specific fields.
Previously it was assuming the argument is a `Channel`, but it is actually a `BaseChannel` (we call it with pipeline.inputs when compiling a PipelineBegin node) so added a branch guard. This fixes the bug that the subpipeline can only take the `OutputChannel`.

Background:

`OutputChannel`, or a `component.outputs["x"]` is not only the possible inputs to another components (i.e. is a `BaseChannel`) but also the place to configure output-related configuration such as `additional_(custom_)properties`, `garbage_collection_policy`, `external_artifact_uris`, etc.

`OutputSpec` is the place to store those output-only information. However there is a case we also accept non-`OutputChannel`; currently only when compiling a PipelineBegin node, which can technically take any `BaseChannel` raw inputs.

Still every node IR needs the most basic `OutputSpec` IR which only stores the output key -> artifact type mapping. They're currently used from the UI to render the pipeline graph.

PiperOrigin-RevId: 482721091