Support concatenating placeholders with itself.
The previous fix using deepcopy would translate `a + '-' + a` into `(a + '-' + a + '-')` as the first concatenation would be included in the deepcopy operation. Instead, we introduce a specific method `concat_placeholders` which is used for concatenating a list of placeholders. At the same time, we disallow concat operations for a placeholder with itself to avoid the problem mentioned previously.

PiperOrigin-RevId: 484086795