Fix KubeFlow entrypoint bug that fails on resolver node post-execution logging.
When Resolver node outputs is used as an inputs, the component IR no longer contains `input_spec.channels`, but the logging was still referencing it for the artifact type retrieval. If this artifact type information is really important we can consider adding a dedicated field to the `InputSpec`. Note that `InputSpec` <-> `BaseChannel` <-> `ArtifactType` has 1:1:1 correspondence.

PiperOrigin-RevId: 484169320