Remove LoopVarChannel.
LoopVarChannel was only required for the old compilation logic. Previously to support `list[dict[str, list[Artifact]]]` type resolver function outputs, we introduced a common interface called `Loopable` (which is just a function that emits with-clause handle), and resolver function returns the loopable. Existing ForEach(single_channel) use case (which is simply an Unnest operation) can also be reduced to the resolver function of list[dict[...]] output type, which eliminates (and simplifies) the LoopVarChannel.

PiperOrigin-RevId: 487452986