Replacing missing features in chicago_taxi_clients.py with their statistical attributes
Using Tensorflow data validation package, the dataset's statistics are calculated. And then when composing a request payload, in case the record is missing a feature, the value is replaced with one the following:

If the feature is of type BYTES (categorical), the most common value will be used instead.
If the feature is numeric, the average value of the column will be used. The average will also be cast to INT if the feature type is integer.