Add `mlmd_state.evict_from_cache` context manager
It is preferable to use `mlmd_execution_atomic_op` for mutating MLMD executions in the orchestrator but sometimes it may be necessary to use third party functions which are not cache aware. Such functions should be invoked within this context manager for proper locking and cache eviction to prevent stale entries.

PiperOrigin-RevId: 494302785