Additional temporarily fixes ImportError on TFX OSS Colab examples by uninstalling shapely
Currently, TFX is supporting 'google-cloud-bigquery>=2.26.0,<3' (so that it installs 2.34.4), which imports from `shapely.geos import WKBWriter`.

However, by default, shapely 2.0.0 is installed on the Colab, that WKBWriter class is deprecated.

This is a temporal solution to avoid an ImportError. Ultimately, it should be handled by supporting a recent version of Bigquery (>=3.0), instead of uninstalling other extra dependencies.

PiperOrigin-RevId: 498302127