Change `get_output_artifacts_from_execution_ids` to use primitive MLMD APIs.
Using MLMD filter query reduces the number of queries (3 -> 2) but the query execution time with bigger `OR` queries seems quite inefficient (taking >100s). Even in general SQL terms, having long OR clause is not an appropriate usage as it harms the optimization of query execution plan.

So we're falling back to primitive graph queries, where the number of total calls can be bigger but total time and the load to the MLMD is smaller.

PiperOrigin-RevId: 517362750