Support list of artifacts in Python custom component.
The input annotated with `InputArtifact[List[MyArtifact]]` will resolve a channel into a list of artifacts form, and the function would be passed a list of artifacts instead of a single artifact enforcement done in normal `InputArtifact[MyArtifact]`. Even if only one artifact is resolved, it would be passed as a list of length 1.

PiperOrigin-RevId: 529141663