Add serialized retry proto message to pass it from component to executor
We do not accept Python class to be serialized into the IR, therefore there was no way to transfer retries.Retry object from the component.

In this change, a new proto message has been defined to encode the retries.Retry, and an utility function that converts it into a Python object has been created.

PiperOrigin-RevId: 530765078