Update the validation to not allow using placeholder input to exec_property
In recent change, we allowed exec_property to take wide placeholder values(except for output & exec property). The syntax is not clear for input, since the component already gets the value from input, it's hard to justify why to reassign input value to exec property. CL adds additional check on this.

PiperOrigin-RevId: 569552046