Remove str from `CWP.__getitem__(index)`
For placeholders in general, it makes sense to accept both int indices (for list-valued placeholders) and string keys (for proto- or JSON-dict-valued placeholders) into `__getitem__()`.

This CL edits the `__get__item()` overrides of ArtifactPlaceholder and ChannelWrappedPlaceholder, where the index is used to select the n-th artifact from the channel. Here, only ints make sense.

PiperOrigin-RevId: 586321783