Implement Placeholder.internal_equals() and use it in CondContext
Because `==` aka. `__eq__` is already reserved for producing predicates, we need a separate function to check if two placeholders are "the same". Note that even non-equal placeholders can eventually _resolve_ to equal values, but that's not what we're concerned with here.

Currently this is marked as `internal_` because Tflex end users are not expected to need this.

PiperOrigin-RevId: 642270043